# Conflicts:
#	python/requirements/ml/requirements_dl.txt
#	python/requirements/ml/requirements_rllib.txt

+++ dashboard/modules/job/job_manager.py
@@ -6,26 +6,34 @@ import os
 import random
 import string
 import subprocess
+import sys
 import time
 import traceback
 from asyncio.tasks import FIRST_COMPLETED
 from collections import deque
 from typing import Any, Dict, Iterator, List, Optional, Tuple
-
+from ray.util.scheduling_strategies import (
+    NodeAffinitySchedulingStrategy,
+    SchedulingStrategyT,
+)
 import ray
 from ray._private.gcs_utils import GcsAioClient
 import ray._private.ray_constants as ray_constants
 from ray._private.runtime_env.constants import RAY_JOB_CONFIG_JSON_ENV_VAR
 from ray.actor import ActorHandle
+from ray.dashboard.consts import RAY_JOB_ALLOW_DRIVER_ON_WORKER_NODES_ENV_VAR
 from ray.dashboard.modules.job.common import (
     JOB_ID_METADATA_KEY,
     JOB_NAME_METADATA_KEY,
+    JOB_ACTOR_NAME_TEMPLATE,
+    SUPERVISOR_ACTOR_RAY_NAMESPACE,
     JobInfo,
     JobInfoStorageClient,
 )
 from ray.dashboard.modules.job.utils import file_tail_iterator
 from ray.exceptions import RuntimeEnvSetupError
 from ray.job_submission import JobStatus
+
 
 logger = logging.getLogger(__name__)
 
@@ -34,6 +42,21 @@ try:
     create_task = asyncio.create_task
 except AttributeError:
     create_task = asyncio.ensure_future
+
+# Windows requires additional packages for proper process control.
+if sys.platform == "win32":
+    try:
+        import win32api
+        import win32con
+        import win32job
+    except (ModuleNotFoundError, ImportError) as e:
+        win32api = None
+        win32con = None
+        win32job = None
+        logger.warning(
+            "Failed to Import win32api. For best usage experience run "
+            f"'conda install pywin32'. Import error: {e}"
+        )
 
 
 def generate_job_id() -> str:
@@ -139,6 +162,9 @@ class JobSupervisor:
         # fire and forget call from outer job manager to this actor
         self._stop_event = asyncio.Event()
 
+        # Windows Job Object used to handle stopping the child processes.
+        self._win32_job_object = None
+
     def _get_driver_runtime_env(self) -> Dict[str, Any]:
         # Get the runtime_env set for the supervisor actor.
         curr_runtime_env = dict(ray.get_runtime_context().runtime_env)
@@ -158,9 +184,13 @@ class JobSupervisor:
         Runs the entrypoint command as a child process, streaming stderr &
         stdout to given log files.
 
+        Unix systems:
         Meanwhile we start a demon process and group driver
         subprocess in same pgid, such that if job actor dies, entire process
         group also fate share with it.
+
+        Windows systems:
+        A jobObject is created to enable fate sharing for the entire process group.
 
         Args:
             logs_path: File path on head node's local disk to store driver
@@ -178,24 +208,58 @@ class JobSupervisor:
                 stderr=subprocess.STDOUT,
             )
             parent_pid = os.getpid()
-            # Create new pgid with new subprocess to execute driver command
             child_pid = child_process.pid
-            child_pgid = os.getpgid(child_pid)
+            # Create new pgid with new subprocess to execute driver command
 
-            # Open a new subprocess to kill the child process when the parent
-            # process dies kill -s 0 parent_pid will succeed if the parent is
-            # alive. If it fails, SIGKILL the child process group and exit
-            subprocess.Popen(
-                f"while kill -s 0 {parent_pid}; do sleep 1; done; kill -9 -{child_pgid}",  # noqa: E501
-                shell=True,
-                # Suppress output
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
+            if sys.platform != "win32":
+                child_pgid = os.getpgid(child_pid)
+
+                # Open a new subprocess to kill the child process when the parent
+                # process dies kill -s 0 parent_pid will succeed if the parent is
+                # alive. If it fails, SIGKILL the child process group and exit
+                subprocess.Popen(
+                    f"while kill -s 0 {parent_pid}; do sleep 1; done; kill -9 -{child_pgid}",  # noqa: E501
+                    shell=True,
+                    # Suppress output
+                    stdout=subprocess.DEVNULL,
+                    stderr=subprocess.DEVNULL,
+                )
+
+            elif sys.platform == "win32" and win32api:
+                # Create a JobObject to which the child process (and its children)
+                # will be connected. This job object can be used to kill the child
+                # processes explicitly or when the jobObject gets deleted during
+                # garbage collection.
+                self._win32_job_object = win32job.CreateJobObject(None, "")
+                win32_job_info = win32job.QueryInformationJobObject(
+                    self._win32_job_object, win32job.JobObjectExtendedLimitInformation
+                )
+                win32_job_info["BasicLimitInformation"][
+                    "LimitFlags"
+                ] = win32job.JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE
+                win32job.SetInformationJobObject(
+                    self._win32_job_object,
+                    win32job.JobObjectExtendedLimitInformation,
+                    win32_job_info,
+                )
+                child_handle = win32api.OpenProcess(
+                    win32con.PROCESS_TERMINATE | win32con.PROCESS_SET_QUOTA,
+                    False,
+                    child_pid,
+                )
+                win32job.AssignProcessToJobObject(self._win32_job_object, child_handle)
+
             return child_process
 
     def _get_driver_env_vars(self) -> Dict[str, str]:
         """Returns environment variables that should be set in the driver."""
+        # RAY_ADDRESS may be the dashboard URL but not the gcs address,
+        # so when the environment variable is not empty, we force set RAY_ADDRESS
+        # to "auto" to avoid function `canonicalize_bootstrap_address_or_die` returning
+        # the wrong GCS address.
+        # TODO(Jialing He, Archit Kulkarni): Definition of Specification RAY_ADDRESS
+        if ray_constants.RAY_ADDRESS_ENVIRONMENT_VARIABLE in os.environ:
+            os.environ[ray_constants.RAY_ADDRESS_ENVIRONMENT_VARIABLE] = "auto"
         ray_addr = ray._private.services.canonicalize_bootstrap_address_or_die(
             "auto", ray.worker._global_node._ray_params.temp_dir
         )
@@ -256,7 +320,21 @@ class JobSupervisor:
             # Block in PENDING state until start signal received.
             await _start_signal_actor.wait.remote()
 
-        await self._job_info_client.put_status(self._job_id, JobStatus.RUNNING)
+        driver_agent_http_address = (
+            "http://"
+            f"{ray.worker.global_worker.node.node_ip_address}:"
+            f"{ray.worker.global_worker.node.dashboard_agent_listen_port}"
+        )
+        driver_node_id = ray.worker.global_worker.current_node_id.hex()
+
+        await self._job_info_client.put_status(
+            self._job_id,
+            JobStatus.RUNNING,
+            jobinfo_replace_kwargs={
+                "driver_agent_http_address": driver_agent_http_address,
+                "driver_node_id": driver_node_id,
+            },
+        )
 
         try:
             # Configure environment variables for the child process. These
@@ -277,8 +355,11 @@ class JobSupervisor:
 
             if self._stop_event.is_set():
                 polling_task.cancel()
-                # TODO (jiaodong): Improve this with SIGTERM then SIGKILL
-                child_process.kill()
+                if sys.platform == "win32" and self._win32_job_object:
+                    win32job.TerminateJobObject(self._win32_job_object, -1)
+                elif sys.platform != "win32":
+                    # TODO (jiaodong): Improve this with SIGTERM then SIGKILL
+                    child_process.kill()
                 await self._job_info_client.put_status(self._job_id, JobStatus.STOPPED)
             else:
                 # Child process finished execution and no stop event is set
@@ -324,9 +405,6 @@ class JobManager:
     goes down.
     """
 
-    JOB_ACTOR_NAME_TEMPLATE = (
-        f"{ray_constants.RAY_INTERNAL_NAMESPACE_PREFIX}job_actor_" + "{job_id}"
-    )
     # Time that we will sleep while tailing logs if no new log line is
     # available.
     LOG_TAIL_SLEEP_S = 1
@@ -354,7 +432,10 @@ class JobManager:
 
     def _get_actor_for_job(self, job_id: str) -> Optional[ActorHandle]:
         try:
-            return ray.get_actor(self.JOB_ACTOR_NAME_TEMPLATE.format(job_id=job_id))
+            return ray.get_actor(
+                JOB_ACTOR_NAME_TEMPLATE.format(job_id=job_id),
+                namespace=SUPERVISOR_ACTOR_RAY_NAMESPACE,
+            )
         except ValueError:  # Ray returns ValueError for nonexistent actor.
             return None
 
@@ -461,6 +542,38 @@ class JobManager:
         runtime_env["env_vars"] = env_vars
         return runtime_env
 
+    async def _get_scheduling_strategy(self) -> SchedulingStrategyT:
+        if os.environ.get(RAY_JOB_ALLOW_DRIVER_ON_WORKER_NODES_ENV_VAR, "0") == "1":
+            logger.info(
+                f"{RAY_JOB_ALLOW_DRIVER_ON_WORKER_NODES_ENV_VAR} was set to 1. "
+                "Using Ray's default actor scheduling strategy for the job "
+                "driver instead of running it on the head node."
+            )
+            scheduling_strategy = "DEFAULT"
+        else:
+            head_node_id_bytes = await self._gcs_aio_client.internal_kv_get(
+                "head_node_id".encode(),
+                namespace=ray_constants.KV_NAMESPACE_JOB,
+                timeout=30,
+            )
+            if head_node_id_bytes is None:
+                logger.info(
+                    "Head node ID not found in GCS. Using Ray's default actor "
+                    "scheduling strategy for the job driver instead of running "
+                    "it on the head node."
+                )
+                scheduling_strategy = "DEFAULT"
+            else:
+                head_node_id = head_node_id_bytes.decode()
+                logger.info(
+                    "Head node ID found in GCS; scheduling job driver on "
+                    f"head node {head_node_id}"
+                )
+                scheduling_strategy = NodeAffinitySchedulingStrategy(
+                    node_id=head_node_id, soft=False
+                )
+        return scheduling_strategy
+
     async def submit_job(
         self,
         *,
@@ -517,16 +630,14 @@ class JobManager:
         # returns immediately and we can catch errors with the actor starting
         # up.
         try:
+            scheduling_strategy = await self._get_scheduling_strategy()
             supervisor = self._supervisor_actor_cls.options(
                 lifetime="detached",
-                name=self.JOB_ACTOR_NAME_TEMPLATE.format(job_id=submission_id),
+                name=JOB_ACTOR_NAME_TEMPLATE.format(job_id=submission_id),
                 num_cpus=0,
-                # Currently we assume JobManager is created by dashboard server
-                # running on headnode, same for job supervisor actors scheduled
-                resources={
-                    self._get_current_node_resource_key(): 0.001,
-                },
+                scheduling_strategy=scheduling_strategy,
                 runtime_env=self._get_supervisor_runtime_env(runtime_env),
+                namespace=SUPERVISOR_ACTOR_RAY_NAMESPACE,
             ).remote(submission_id, entrypoint, metadata or {}, self._gcs_address)
             supervisor.run.remote(_start_signal_actor=_start_signal_actor)
 
@@ -556,6 +667,9 @@ class JobManager:
         else:
             return False
 
+    def job_info_client(self) -> JobInfoStorageClient:
+        return self._job_info_client
+
     async def get_job_status(self, job_id: str) -> Optional[JobStatus]:
         """Get latest status of a job."""
         return await self._job_info_client.get_status(job_id)
@@ -581,7 +695,7 @@ class JobManager:
             if lines is None:
                 # Return if the job has exited and there are no new log lines.
                 status = await self.get_job_status(job_id)
-                if status not in {JobStatus.PENDING, JobStatus.RUNNING}:
+                if status.is_terminal():
                     return
 
                 await asyncio.sleep(self.LOG_TAIL_SLEEP_S)
